more stuff probably lost in merge: fix a bunch of bogus casts between pointers and long

+++ Engine/gui/guidialog.cpp
@@ -192,11 +192,11 @@ int savegamedialog()
 
   lpTemp = nullptr;
   if (numsaves > 0)
-    CSCISendControlMessage(ctrllist, CLB_GETTEXT, 0, (long)&buffer2[0]);
+    CSCISendControlMessage(ctrllist, CLB_GETTEXT, 0, &buffer2[0]);
   else
     buffer2[0] = 0;
 
-  CSCISendControlMessage(ctrltbox, CTB_SETTEXT, 0, (long)&buffer2[0]);
+  CSCISendControlMessage(ctrltbox, CTB_SETTEXT, 0, &buffer2[0]);
 
   int toret = -1;
   while (1) {
@@ -204,10 +204,10 @@ int savegamedialog()
     if (mes.code == CM_COMMAND) {
       if (mes.id == ctrlok) {
         int cursell = CSCISendControlMessage(ctrllist, CLB_GETCURSEL, 0, 0);
-        CSCISendControlMessage(ctrltbox, CTB_GETTEXT, 0, (long)&buffer2[0]);
+        CSCISendControlMessage(ctrltbox, CTB_GETTEXT, 0, &buffer2[0]);
 
         if (numsaves > 0)
-          CSCISendControlMessage(ctrllist, CLB_GETTEXT, cursell, (long)&bufTemp[0]);
+          CSCISendControlMessage(ctrllist, CLB_GETTEXT, cursell, &bufTemp[0]);
         else
           strcpy(bufTemp, "_NOSAVEGAMENAME");
 
@@ -231,7 +231,7 @@ int savegamedialog()
             CSCIWaitMessage(&cmes);
           } while (cmes.code != CM_COMMAND);
 
-          CSCISendControlMessage(txt1, CTB_GETTEXT, 0, (long)&buffer2[0]);
+          CSCISendControlMessage(txt1, CTB_GETTEXT, 0, &buffer2[0]);
           CSCIDeleteControl(btnCancel);
           CSCIDeleteControl(btnOk);
           CSCIDeleteControl(txt1);
@@ -282,8 +282,8 @@ int savegamedialog()
     } else if (mes.code == CM_SELCHANGE) {
       int cursel = CSCISendControlMessage(ctrllist, CLB_GETCURSEL, 0, 0);
       if (cursel >= 0) {
-        CSCISendControlMessage(ctrllist, CLB_GETTEXT, cursel, (long)&buffer2[0]);
-        CSCISendControlMessage(ctrltbox, CTB_SETTEXT, 0, (long)&buffer2[0]);
+        CSCISendControlMessage(ctrllist, CLB_GETTEXT, cursel, &buffer2[0]);
+        CSCISendControlMessage(ctrltbox, CTB_SETTEXT, 0, &buffer2[0]);
       }
     }
   }
@@ -391,7 +391,7 @@ void enterstringwindow(const char *prompttext, char *stouse)
       if (mes.id == ctrlcancel)
         buffer2[0] = 0;
       else
-        CSCISendControlMessage(ctrltbox, CTB_GETTEXT, 0, (long)&buffer2[0]);
+        CSCISendControlMessage(ctrltbox, CTB_GETTEXT, 0, &buffer2[0]);
       break;
     }
   }
@@ -433,7 +433,7 @@ int roomSelectorWindow(int currentRoom, int numRooms, int*roomNumbers, char**roo
   for (int aa = 0; aa < numRooms; aa++)
   {
     sprintf(buff, "%3d %s", roomNumbers[aa], roomNames[aa]);
-    CSCISendControlMessage(ctrllist, CLB_ADDITEM, 0, (long)&buff[0]);
+    CSCISendControlMessage(ctrllist, CLB_ADDITEM, 0, &buff[0]);
     if (roomNumbers[aa] == currentRoom)
     {
       CSCISendControlMessage(ctrllist, CLB_SETCURSEL, aa, 0);
@@ -448,7 +448,7 @@ int roomSelectorWindow(int currentRoom, int numRooms, int*roomNumbers, char**roo
   buffer2[0] = 0;
 
   int ctrltbox = CSCICreateControl(CNT_TEXTBOX, 10, 29, 120, 0, nullptr);
-  CSCISendControlMessage(ctrltbox, CTB_SETTEXT, 0, (long)&buffer2[0]);
+  CSCISendControlMessage(ctrltbox, CTB_SETTEXT, 0, &buffer2[0]);
 
   int toret = -1;
   while (1) {
@@ -457,7 +457,7 @@ int roomSelectorWindow(int currentRoom, int numRooms, int*roomNumbers, char**roo
     {
       if (mes.id == ctrlok) 
       {
-        CSCISendControlMessage(ctrltbox, CTB_GETTEXT, 0, (long)&buffer2[0]);
+        CSCISendControlMessage(ctrltbox, CTB_GETTEXT, 0, &buffer2[0]);
         if (std::isdigit(buffer2[0]))
         {
           toret = atoi(buffer2);
@@ -474,7 +474,7 @@ int roomSelectorWindow(int currentRoom, int numRooms, int*roomNumbers, char**roo
       if (cursel >= 0) 
       {
         sprintf(buffer2, "%d", roomNumbers[cursel]);
-        CSCISendControlMessage(ctrltbox, CTB_SETTEXT, 0, (long)&buffer2[0]);
+        CSCISendControlMessage(ctrltbox, CTB_SETTEXT, 0, &buffer2[0]);
       }
     }
   }
